Rename: mockData file path
